requirements.txt
django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.7
Pillow==9.5.0